% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_write_r.R
\name{i_write_r}
\alias{i_write_r}
\title{I Write R}
\usage{
i_write_r(number, noun)
}
\arguments{
\item{noun}{}
}
\value{

}
\description{
This function tells you a bit about yourself given your prompted input
}
