% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pizza_maker.R
\name{pizza_maker}
\alias{pizza_maker}
\title{Pizza Maker}
\usage{
pizza_maker(num_pizzas, cheese_type, topping)
}
\arguments{
\item{num_pizzas}{a number indicating the amount of pizzas to make}

\item{cheese_type}{a character containing the type of cheese to use}

\item{topping}{a character containing the topping of your choice}
}
\value{

}
\description{
This is a very stupid function that makes any number of pizzas with a cheese and topping of your choice
}
