% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_super_rad_butnotrly.R
\name{theme_super_rad_butnotrly}
\alias{theme_super_rad_butnotrly}
\title{A theme falsely claiming to be super rad, but then admitting to that falsehood}
\usage{
theme_super_rad_butnotrly()
}
\value{

}
\description{
don't get your hopes up
}
